Atualização e tradução das classes
